Added typedefs to eliminate MSVC6 warnings.

+++ src/sbbs3/xtrn_sec.cpp
@@ -8,7 +8,7 @@
  * @format.tab-size 4		(Plain Text/Source Code File Header)			*
  * @format.use-tabs true	(see http://www.synchro.net/ptsc_hdr.html)		*
  *																			*
- * Copyright 2006 Rob Swindell - http://www.synchro.net/copyright.html		*
+ * Copyright 2007 Rob Swindell - http://www.synchro.net/copyright.html		*
  *																			*
  * This program is free software; you can redistribute it and/or			*
  * modify it under the terms of the GNU General Public License				*
@@ -817,7 +817,7 @@ void sbbs_t::xtrndat(char *name, char *dropdir, uchar type, ulong tleft
 		write(file,&c,1);						/* GraphicsMode */
 		c=useron.xedit ? 1:0;
 		write(file,&c,1);						/* ExternEdit */
-		i=rows;
+		i=(int16_t)rows;
 		write(file,&i,2);						/* ScreenLength */
 		c=1;
 		write(file,&c,1);						/* MNP_Connect */
@@ -1008,7 +1008,7 @@ void sbbs_t::xtrndat(char *name, char *dropdir, uchar type, ulong tleft
 		sprintf(str,"%-25.25s",name);           /* User's full name */
 		write(file,str,25);
 
-		i=(tleft/60);
+		i=(int16_t)(tleft/60);
 		write(file,&i,2);						/* Minutes remaining */
 
 		write(file,&cfg.node_num,1);			/* Node number */
@@ -1361,7 +1361,7 @@ void sbbs_t::moduserdat(uint xtrnnum)
 			lseek(file,373,SEEK_SET);
 			read(file,&i,2);			/* SecLvl */
 			if(i<SYSOP_LEVEL) {
-				useron.level=i;
+				useron.level=(uint8_t)i;
 				putuserrec(&cfg,useron.number,U_LEVEL,2,ultoa(useron.level,tmp,10)); 
 			}
 			close(file);
@@ -1448,7 +1448,7 @@ void sbbs_t::moduserdat(uint xtrnnum)
 				lseek(file,105,SEEK_CUR);	/* read security level */
 				read(file,&i,2);
 				if(i<SYSOP_LEVEL) {
-					useron.level=i;
+					useron.level=(uint8_t)i;
 					putuserrec(&cfg,useron.number,U_LEVEL,2,ultoa(useron.level,tmp,10)); 
 				}
 				lseek(file,75,SEEK_CUR);	/* read in expiration date */
